+++ lib/internal/Magento/Framework/Stdlib/DateTime/DateTime.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright © 2015 Magento. All rights reserved.
+ * Copyright © 2016 Magento. All rights reserved.
  * See COPYING.txt for license details.
  */
 
@@ -64,8 +64,6 @@ class DateTime
      * @param  string $format
      * @param  int|string $input date in current timezone
      * @return string
-     *
-     * @deprecated (MAGETWO-35555)
      */
     public function gmtDate($format = null, $input = null)
     {
@@ -107,7 +105,7 @@ class DateTime
     public function gmtTimestamp($input = null)
     {
         if ($input === null) {
-            return gmdate('U');
+            return (int)gmdate('U');
         } elseif (is_numeric($input)) {
             $result = $input;
         } else {
